feat: Add Transformers.js chat implementation
This commit adds support for using Transformers.js as an alternative to Wllama for the chat functionality.

A selection page is added at `/chat` to allow you to choose between the two implementations.

- The Wllama implementation is moved to `/chat/wllama`.
- The new Transformers.js implementation is at `/chat/transformers`.
- The configuration has been generalized to support both.
