+++ package.json
@@ -42,8 +42,9 @@
 		"vitest": "^3.2.3"
 	},
 	"dependencies": {
-		"express": "^5.1.0",
+		"@fontsource/bebas-neue": "^5.0.20",
 		"@fontsource/space-grotesk": "^5.0.20",
-		"@fontsource/bebas-neue": "^5.0.20"
+		"@xenova/transformers": "^2.17.2",
+		"express": "^5.1.0"
 	}
 }
