+++ src/lib/chat-config.ts
@@ -1,12 +1,3 @@
-// Configuration for Wllama
-import singleThreaded from '@wllama/wllama/src/single-thread/wllama.wasm?url';
-import multiThreaded from '@wllama/wllama/src/multi-thread/wllama.wasm?url';
-
-export const WLLAMA_CONFIG_PATHS = {
-	'single-thread/wllama.wasm': singleThreaded,
-	'multi-thread/wllama.wasm': multiThreaded
-};
-
 export const DEFAULT_INFERENCE_PARAMS = {
 	nThreads: -1, // auto
 	nContext: 4096,
@@ -18,8 +9,8 @@ export const DEFAULT_INFERENCE_PARAMS = {
 export const DEFAULT_CHAT_TEMPLATE =
 	"{% for message in messages %}{{'<|im_start|>' + message['role'] + '\n' + message['content'] + '<|im_end|>' + '\n'}}{% endfor %}{% if add_generation_prompt %}{{ '<|im_start|>assistant\n' }}{% endif %}";
 
-// Models available for download
-export const AVAILABLE_MODELS = [
+// Models available for Wllama
+export const WLLAMA_MODELS = [
 	{
 		url: 'https://files.khromov.se/models/google_gemma-3-1b-it-qat-Q4_0.gguf',
 		name: 'Gemma3 1B',
@@ -33,6 +24,14 @@ export const AVAILABLE_MODELS = [
 		size: 386404992
 	}
 ];
+
+// Models available for Transformers.js
+export const TRANSFORMERS_MODELS = [
+    {
+        id: "Xenova/llama-3.2-8b-it-gguf",
+        name: "Llama 3.2 8B",
+    }
+]
 
 export interface Message {
 	role: 'user' | 'assistant' | 'system';
